requirements.txt
asgiref==3.3.4
Django==3.2
django-materializecss-form==1.1.17
mysqlclient==2.0.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
